Update coding keyboard layout
A bunch of keys on the main layer have been moved around, and the layer
for mouse navigation is now also used for the arrow keys. Most notably,
I moved the V to the key that normally is used for a semicolon. This
frees up the old position to use for for switching to layer 2. This is
easier than the old approach where I had to press two buttons to switch
to this layer.